Add files via upload
These are what I have so far in terms of cleaning the dataset. Some weird things are happening with the 0's. I do not know if we want the 0s in our dataset, as they just mean that the country did not have any data on it at the time. It kind of skews our stats if we keep the 0s.